Refactor economic model implementation and enhance reporting capabilities
- Introduced a new EconomicModel class to replace the previous Model class, streamlining the initialization process and improving parameter management.
- Updated tests to utilize the new EconomicModel class, ensuring accurate calculations and validations.
- Enhanced markdown report generation to include detailed sections for parameters, results, assumptions, and limitations, improving clarity and usability.
- Removed outdated model and type files to clean up the codebase and focus on the new structure.

These changes improve the maintainability and usability of the economic modeling framework, facilitating better analysis and reporting of economic metrics.
