--- packages/mathematical-modeling/src/model.ts
@@ -1,1 +1,0 @@
-export { BaseModel } from './types';
